@@ -42,7 +42,7 @@ async function setJSONValues() {
       answers = result.answers;
     }
   } catch (err) {
-    console.log('An error occurred while getting questions/answers/practiceQuestions: ' + err.message);
+    console.log(`An error occurred while getting questions/answers/practiceQuestions: ${err.message}`);
   }
 }
 
@@ -71,7 +71,7 @@ router.get(URI.DATE.CAN_START, async (req: Request, res: Response) => {
       res.json(false);
     }
   } catch (err) {
-    console.log('An error occurred while getting questions/answers/practiceQuestions: ' + err.message);
+    console.log(`An error occurred while getting questions/answers/practiceQuestions: ${err.message}`);
     res.set(500).end();
   }
 });
@@ -88,7 +88,7 @@ router.post(URI.TEAM.SAVE, async (req: Request, res: Response) => {
     const result = await collection.insertOne(team);
     res.json(result.ops[0]);
   } catch (err) {
-    console.log('An error occurred while updating a team: ' + err.message);
+    console.log(`An error occurred while updating a team: ${err.message}`);
     res.set(500).end();
   }
 });
@@ -107,7 +107,7 @@ router.put(URI.TEAM.SAVE, async (req: Request, res: Response) => {
     });
     res.json(team);
   } catch (err) {
-    console.log('An error occurred while saving a team: ' + err.message);
+    console.log(`An error occurred while saving a team: ${err.message}`);
     res.set(500).end();
   }
 });
@@ -125,18 +125,18 @@ router.put(URI.TEAM.GET, async (req: Request, res: Response) => {
       res.json(result);
     }
   } catch (err) {
-    console.log('An error occurred while getting a team: ' + err.message);
+    console.log(`An error occurred while getting a team: ${err.message}`);
     res.set(500).end();
   }
 });
 
 router.post(URI.ADMIN.LOGIN, (req: Request, res: Response) => {
-  const password = req.body.password;
+  const password = req.body;
   if (password === adminPassword) {
     const payload = { type: 'admin', check:  true };
     const options: SignOptions = { expiresIn: 1440 }; // expires in 24 hours
     const token = jwtSign(payload, secret, options);
-    res.json({token});
+    res.send(token);
   } else {
     res.set(403).end();
   }
